backend/src/requirements.txt
fastapi==0.54.1
psycopg2-binary==2.8.5
SQLAlchemy==1.3.16
uvicorn==0.11.3
passlib[bcrypt]
python-multipart
pyjwt
pydantic[email]
pymongo

# dev
black
mypy
flake8
requests
pytest
pytest-cov
isort
autoflake